+++ src/features/adr/components/index.ts
@@ -1,0 +1,3 @@
+export { PatientDetails } from "./PatientDetails";
+export { PatientList } from "./PatientList";
+export { PatientLookup } from "./PatientLookup";
